Add files via upload
Preliminary code for the phase shift function on the arduino board. Takes in voltage value as argument and converts the voltage value to a binary stream for the pins to output to the phase shift module. in the code it is only written for one ananlog pin and 6 digital pins but code would just have to be run in a loop to get all the pinouts to return digital data to the phase shifters.